init sidebar config in _app.page.tsx
